Add PCA data support to data prep reader (#50)
## Summary
- refactor data prep readers into unified runner hierarchy
- add PcaRunner and consolidate tank/PCA reading into single reader
script
- update configs and scripts for new `pca` type

## Testing
- `pytest`
- `python -m py_compile data_prep/src/runner.py data_prep/src/reader.py
data_prep/src/config.py data_prep/scripts/gen_data.py`


------
https://chatgpt.com/codex/tasks/task_e_68a346d0d4308325940cb1724b9fcf15
