--- data_prep/src/data_reader/__init__.py
@@ -1,3 +1,0 @@
-from data_reader.config import Config  # noqa: F401
-from data_reader.run import run  # noqa: F401
-from data_reader.run import Parameters  # noqa: F401
