+++ data_prep/src/reader.py
@@ -1,23 +1,167 @@
+import os
+import sys
+import argparse
+from datetime import datetime, timedelta
+import numpy as np
+import bufr
 
-from datetime import datetime
+sys.path.insert(0, os.path.realpath('/'))
+from zarr_encoder import Encoder as ZarrEncoder  # noqa: E402
+from parquet_encoder import Encoder as ParquetEncoder  # noqa: E402
+import runner  # noqa: E402
+import settings  # noqa: E402
 
-class Reader:
-    def __init__(self, data_path: str):
-        pass
 
-    def create_data(start_date: datetime,
-                    end_date: datetime,
-                    data_type: str,
-                    suffix: str = None,
-                    append: bool = True) -> None:
-        assert False, "This method should be implemented in a subclass."
+def create_data(start_date: datetime,
+                end_date: datetime,
+                data_type: str,
+                output_type: str,
+                suffix: str = None,
+                append: bool = True) -> None:
+    """Create zarr files from BUFR data in week long chunks."""
 
-class ReaderFactory:
-    @staticmethod
-    def create_reader(data_type: str, data_path: str):
-        if data_type == 'bufr':
-            from .bufr_pca_reader import Reader
-            return Reader(data_path)
+    bufr.mpi.App(sys.argv)
+    comm = bufr.mpi.Comm("world")
+
+    extension = 'zarr' if output_type == 'zarr' else 'pqt'
+
+    if suffix:
+        file_name = f"{data_type}_{suffix}.{extension}"
+    else:
+        file_name = f"{data_type}.{extension}"
+
+    output_path = os.path.join(settings.OUTPUT_PATH, file_name)
+
+    if output_type == 'zarr':
+        if comm.rank() == 0:
+            # Ensure all output directories exist before processing
+            if not append and os.path.exists(output_path):
+                import shutil
+                shutil.rmtree(output_path)
+            os.makedirs(output_path, exist_ok=True)
+
+    comm.barrier()
+
+    date = start_date
+    day = timedelta(days=1)
+
+    while date <= end_date:
+        _create_data_for_day(comm, date, data_type, output_type, output_path)
+        date += day
+
+def create_weekly_data(start_date: datetime,
+                       end_date: datetime,
+                       data_type: str,
+                       output_type: str = 'parquet',
+                       suffix: str = None,
+                       append: bool = True) -> None:
+    """Create zarr files from BUFR data in week long chunks."""
+
+    bufr.mpi.App(sys.argv)
+    comm = bufr.mpi.Comm("world")
+
+    # Determine all week boundaries (Monday - Sunday) that intersect the range
+    week_start = start_date - timedelta(days=start_date.weekday())
+    week_ranges = []
+    while week_start <= end_date:
+        week_end = week_start + timedelta(days=6)
+        week_ranges.append((week_start, week_end))
+        week_start = week_end + timedelta(days=1)
+
+    extension = 'zarr' if output_type == 'zarr' else 'pqt'
+
+    # Generate output paths for each week
+    output_paths = {}
+    for wstart, wend in week_ranges:
+        if suffix:
+            file_name = f"{data_type}_{suffix}_{wstart:%Y%m%d}_{wend:%Y%m%d}.{extension}"
         else:
-            raise ValueError(f"Unsupported data type: {data_type}")
+            file_name = f"{data_type}_{wstart:%Y%m%d}_{wend:%Y%m%d}.{extension}"
+        output_paths[(wstart, wend)] = os.path.join(settings.OUTPUT_PATH, file_name)
 
+    if output_type == 'zarr' and comm.rank() == 0:
+        # Ensure all output directories exist before processing
+        for path in output_paths.values():
+            if not append and os.path.exists(path):
+                import shutil
+                shutil.rmtree(path)
+            os.makedirs(path, exist_ok=True)
+        comm.barrier()
+
+    # Process each day and append to the appropriate weekly file
+    day = timedelta(days=1)
+    date = start_date
+    while date <= end_date:
+        week_start = date - timedelta(days=date.weekday())
+        week_end = week_start + timedelta(days=6)
+        out_path = output_paths[(week_start, week_end)]
+
+        _create_data_for_day(comm, date, data_type, output_type, out_path)
+        date += day
+
+
+def _create_data_for_day(comm,
+                        date: datetime,
+                        data_type: str,
+                        output_type: str,
+                        output_path: str,
+                        append: bool = True) -> None:
+    start_datetime = date
+    end_datetime = date + timedelta(hours=23, minutes=59, seconds=59)
+
+    parameters = runner.Parameters()
+    parameters.start_time = start_datetime
+    parameters.stop_time = end_datetime
+
+    description, container = runner.run(comm, data_type, parameters)
+
+    if container is None:
+        raise ValueError("No data found")
+
+    # Filter data based on the specified latitude and longitude ranges
+    # if the settings have been defined
+    if hasattr(settings, 'LAT_RANGE') and hasattr(settings, 'LON_RANGE'):
+        latitudes = container.get('latitude')
+        longitudes = container.get('longitude')
+
+        mask = np.array([True] * len(latitudes))
+        mask[latitudes < settings.LAT_RANGE[0]] = False
+        mask[latitudes > settings.LAT_RANGE[1]] = False
+        mask[longitudes < settings.LON_RANGE[0]] = False
+        mask[longitudes > settings.LON_RANGE[1]] = False
+
+        if not np.any(mask):
+            return  # No data in the region
+
+        container.apply_mask(mask)
+
+    if comm.rank() == 0:
+        if output_type == 'zarr':
+            ZarrEncoder(description).encode(container, f'{output_path}', append=append)
+        elif output_type == 'parquet':
+            ParquetEncoder(description).encode(container, f'{output_path}', append=append)
+        else:
+            raise ValueError(f"Unsupported output type: {output_type}")
+
+        print(f"Output written to {output_path}")
+        sys.stdout.flush()
+
+
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser()
+    parser.add_argument('start_date')
+    parser.add_argument('end_date')
+    parser.add_argument('type')
+    parser.add_argument('output_type', choices=['zarr', 'parquet'], help='Output file type')
+    parser.add_argument('-s', '--suffix', required=False, help='Suffix for the output file(s)')
+    parser.add_argument('-a', '--append', action='store_true', help='Append to existing data')
+
+    args = parser.parse_args()
+
+    start_date = datetime.strptime(args.start_date, "%Y-%m-%d")
+    end_date = datetime.strptime(args.end_date, "%Y-%m-%d")
+
+    if args.output_type == 'zarr':
+        create_data(start_date, end_date, args.type, args.output_type, args.suffix, args.append)
+    elif args.output_type == 'parquet':
+        create_weekly_data(start_date, end_date, args.type, args.output_type, args.suffix, args.append)
